gdb/darwin: skip over WIFSTOPPED wait4 status
On modern Darwin's, there appears to be a new circumstance in which a
MACH_NOTIFY_DEAD_NAME message can be received, and which was not
previously accounted for: to signal the WIFSTOPPED condition in the
debuggee. In that case the debuggee is not dead yet (and in fact,
counting it as dead would cause a zombie leak - A process in such a
state reparents to PID 1, but cannot be killed).

- Read and ignore such messages (counting on the next exception
message to let us know of the inferior's new state again)
- Refactor logging so as to clearly distinguish between the three
MACH_NOTIFY_DEAD_NAME cases (WIFEXITED, WIFSTOPPED, signal)

Co-authored-by: Louis-He <1726110778@qq.com>
Co-authored-by: Philippe Blain <levraiphilippeblain@gmail.com>

ChangeLog:

2022-02-05  Dominique Quatravaux <dominique.quatravaux@epfl.ch>

	PR gdb/24609
	* gdb/darwin-nat.c (darwin_nat_target::decode_message): Also
	check for WIFSTOPPED upon MACH_NOTIFY_DEAD_NAME.